Extract renderToHTMLWithComponents
